🚚 Rename: activity-detail ui 폴더 구조 정리
