+++ src/features/activity/activity-detail/ui/activity-info/ActivityDescription.tsx
