+++ src/features/activity/activity-detail/ui/activity-info/ActivityInfo.tsx
