+++ src/features/activity/activity-detail/ui/activity-info/ImageCarousel.tsx
