+++ src/features/activity/activity-detail/ui/tab/ActivityTabSection.tsx
@@ -3,8 +3,8 @@
 import { useEffect, useRef, useState } from 'react';
 import TabBar from '@/shared/ui/tab-bar/TabBar';
 import { cn } from '@/shared/utils/cn';
-import ActivityDescription from './ActivityDescription';
-import ActivityLocation from './ActivityLocation';
+import ActivityDescription from '../activity-info/ActivityDescription';
+import ActivityLocation from '../activity-info/ActivityLocation';
 
 type ActivityTabSectionProps = {
   description: string;
